Moving devOptions to new folder Dev
