+++ src/components/Main/index.ts
@@ -1,5 +1,4 @@
-import AppMenu from './AppMenu'
 import Footer from './Footer'
 import ModalWarning from './ModalWarning'
 
-export { AppMenu, Footer, ModalWarning }
+export { Footer, ModalWarning }
