+++ src/components/index.ts
@@ -13,7 +13,7 @@ import FilterList from './FilterList'
 import Home from './Home'
 import Login from '../containers/User/Login'
 import Loading from './Loading'
-import { AppMenu, Footer, ModalWarning } from './Main'
+import { Footer, ModalWarning } from './Main'
 import { DecisionArea, ReviewSection } from './Review'
 import TemplateList from './TemplateList'
 import TemplateNew from './TemplateNew'
@@ -29,7 +29,6 @@ export {
   Account,
   AdminPermissions,
   AdminUsers,
-  AppMenu,
   ApplicationEdit,
   ApplicationHeader,
   ApplicationSelectType,
