+++ src/containers/Dev/AppMenu.tsx
@@ -3,7 +3,7 @@ import { Button, Grid, Icon, Popup } from 'semantic-ui-react'
 import { Link, RouteComponentProps, withRouter } from 'react-router-dom'
 import { useRouter } from '../../utils/hooks/useRouter'
 import { TemplatesDetails } from '../../utils/types'
-import Loading from '../Loading'
+import Loading from '../../components/Loading'
 
 interface AppMenuProps extends RouteComponentProps {
   templatePermissions: TemplatesDetails
